Move test_utils.ts to tests/utils/test_utils.ts
