+++ renderers/angular/a2ui_explorer/src/app/tests/v0_9/04_weather-current.spec.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {getCanvas, loadExample} from '../test_utils';
+import {getCanvas, loadExample} from '../utils';
 
 describe('Example: Weather Current', () => {
   let textContent: string;
