+++ renderers/angular/a2ui_explorer/src/app/tests/v0_9/16_workout-summary.spec.ts
@@ -16,7 +16,7 @@
 
 import {ComponentFixture} from '@angular/core/testing';
 import {DemoComponent} from '../../demo.component';
-import {getCanvas, loadExample} from '../test_utils';
+import {getCanvas, loadExample} from '../utils';
 
 describe('Example: Workout Summary', () => {
   let fixture: ComponentFixture<DemoComponent>;
