+++ renderers/angular/a2ui_explorer/src/app/tests/v0_9/19_software-purchase.spec.ts
@@ -16,7 +16,7 @@
 
 import {ComponentFixture} from '@angular/core/testing';
 import {DemoComponent} from '../../demo.component';
-import {getCanvas, loadExample, wait} from '../test_utils';
+import {getCanvas, loadExample, wait} from '../utils';
 
 describe('Example: Software Purchase', () => {
   let fixture: ComponentFixture<DemoComponent>;
