+++ renderers/angular/a2ui_explorer/src/app/tests/v0_9/28_countdown-timer.spec.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {getCanvas, loadExample} from '../test_utils';
+import {getCanvas, loadExample} from '../utils';
 
 describe('Example: Countdown Timer', () => {
   let textContent: string;
