+++ renderers/angular/a2ui_explorer/src/app/tests/v0_9/34_child-list-template.spec.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import {getCanvas, loadExample} from '../test_utils';
+import {getCanvas, loadExample} from '../utils';
 
 describe('Example: ChildList Template Expansion', () => {
   let textContent: string;
