+++ renderers/angular/a2ui_explorer/tsconfig.app.json
@@ -8,5 +8,5 @@
     "moduleResolution": "bundler"
   },
   "include": ["src/**/*.ts"],
-  "exclude": ["src/**/*.spec.ts", "src/app/tests/test_utils.ts"]
+  "exclude": ["src/**/*.spec.ts", "src/app/tests/utils/**"]
 }
